// src/generated/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
// ...
export interface Scalars {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  DateTime: { input: any; output: any; }
  JSONObject: { input: any; output: any; }
  Time: { input: any; output: any; }
}

export interface ApiError extends Error {
  __typename?: 'ApiError';
  args?: Maybe<Scalars['JSONObject']['output']>;
  code?: Maybe<Scalars['Int']['output']>;
  data?: Maybe<Scalars['JSONObject']['output']>;
  message?: Maybe<Scalars['String']['output']>;
  path?: Maybe<Scalars['String']['output']>;
}

export interface CreateHost {
  deviceKey: Scalars['String']['input'];
  deviceType: Scalars['String']['input'];
  /**
   * groupNames is used to assign the created object
   * to a host group. It is mandatory but
   * can also be blank. This is usefull in case of
   * passing a groupid instead which is
   * the zabbix internal key for storing the group.
   * If a groupid is provided the passed groupName is ignored
   */
  groupNames: Array<Scalars['String']['input']>;
  /**
   * Optionally the internal groupids can be passed - in this case the
   * groupName is ignored
   */
  groupids?: InputMaybe<Array<InputMaybe<Scalars['Int']['input']>>>;
  location?: InputMaybe<LocationInput>;
  /** Optional display name of the device (must be unique if provided - default is to set display name to deviceKey) */
  name?: InputMaybe<Scalars['String']['input']>;
}

export interface CreateHostGroup {
  /** Name of the host group */
  groupName: Scalars['String']['input'];
  /**
   * Internally used unique id
   * (will be assigned by Zabbix if empty)
   */
  uuid?: InputMaybe<Scalars['String']['input']>;
}

export interface CreateHostGroupResponse {
  __typename?: 'CreateHostGroupResponse';
  error?: Maybe<ApiError>;
  groupName: Scalars['String']['output'];
  groupid?: Maybe<Scalars['Int']['output']>;
  message?: Maybe<Scalars['String']['output']>;
}

export interface CreateHostResponse {
  __typename?: 'CreateHostResponse';
  deviceKey: Scalars['String']['output'];
  error?: Maybe<ApiError>;
  hostid?: Maybe<Scalars['String']['output']>;
  message?: Maybe<Scalars['String']['output']>;
}

/**
 * (IoT / Edge - ) Devices are hosts having a state containing the "output" / the business data which is exposed
 * besides monitoring information.
 */
export interface Device {
  deviceType?: Maybe<Scalars['String']['output']>;
  /** Per convention a uuid is used as hostname to identify devices if they do not have a unique hostname */
  host: Scalars['String']['output'];
  hostgroups?: Maybe<Array<HostGroup>>;
  hostid: Scalars['ID']['output'];
  name?: Maybe<Scalars['String']['output']>;
  state?: Maybe<DeviceState>;
  tags?: Maybe<Scalars['JSONObject']['output']>;
}

export { DeviceCommunicationType };

export interface DeviceState {
  operational?: Maybe<OperationalDeviceData>;
}

export { DeviceStatus };

export interface Error {
  code?: Maybe<Scalars['Int']['output']>;
  data?: Maybe<Scalars['JSONObject']['output']>;
  message?: Maybe<Scalars['String']['output']>;
}

export interface ErrorPayload {
  __typename?: 'ErrorPayload';
  additionalInfo?: Maybe<Scalars['JSONObject']['output']>;
  code: Scalars['Int']['output'];
  message?: Maybe<Scalars['String']['output']>;
}
// ...
export interface Host {
  deviceType?: Maybe<Scalars['String']['output']>;
  /** The host field contains the "hostname" in Zabbix */
  host: Scalars['String']['output'];
  hostgroups?: Maybe<Array<HostGroup>>;
  hostid: Scalars['ID']['output'];
  name?: Maybe<Scalars['String']['output']>;
  tags?: Maybe<Scalars['JSONObject']['output']>;
}
// ...
export interface LocationInput {
  location_lat?: InputMaybe<Scalars['String']['input']>;
  location_lon?: InputMaybe<Scalars['String']['input']>;
  name?: InputMaybe<Scalars['String']['input']>;
}
// ...
export interface MutationCreateHostArgs {
  host: Scalars['String']['input'];
  hostgroupids: Array<Scalars['Int']['input']>;
  location?: InputMaybe<LocationInput>;
  templateids: Array<Scalars['Int']['input']>;
}


export interface MutationImportHostGroupsArgs {
  hostGroups: Array<CreateHostGroup>;
}


export interface MutationImportHostsArgs {
  hosts: Array<CreateHost>;
}


export interface MutationImportUserRightsArgs {
  dryRun?: Scalars['Boolean']['input'];
  input: UserRightsInput;
}

export interface OperationalDeviceData {
  __typename?: 'OperationalDeviceData';
  error?: Maybe<Array<ErrorPayload>>;
  location?: Maybe<Location>;
  signalstrength?: Maybe<Scalars['Float']['output']>;
  temperature?: Maybe<Scalars['Float']['output']>;
  timestamp?: Maybe<Scalars['DateTime']['output']>;
  voltage?: Maybe<Scalars['Float']['output']>;
}

export { Permission };

export interface PermissionRequest {
  /**
   * objectName maps to name / path suffix of the template group representing the permission in Zabbix:
   * Permissions/{objectName}
   */
  objectName: Scalars['String']['input'];
  permission: Permission;
}
// ...
export interface QueryAllHostGroupsArgs {
  search_name?: InputMaybe<Scalars['String']['input']>;
  with_hosts?: InputMaybe<Scalars['Boolean']['input']>;
}


export interface QueryAllHostsArgs {
  filter_host?: InputMaybe<Scalars['String']['input']>;
  groupids?: InputMaybe<Array<Scalars['Int']['input']>>;
  hostids?: InputMaybe<Scalars['Int']['input']>;
  name_pattern?: InputMaybe<Scalars['String']['input']>;
  tag_deviceType?: InputMaybe<Array<InputMaybe<Scalars['String']['input']>>>;
  tag_hostType?: InputMaybe<Array<Scalars['String']['input']>>;
  with_items?: InputMaybe<Scalars['Boolean']['input']>;
}
// ...
export interface QueryExportUserRightsArgs {
  exclude_hostgroups_pattern?: InputMaybe<Scalars['String']['input']>;
  name_pattern?: InputMaybe<Scalars['String']['input']>;
}


export interface QueryHasPermissionsArgs {
  permissions: Array<PermissionRequest>;
}


export interface QueryLocationsArgs {
  distinct_by_name?: InputMaybe<Scalars['Boolean']['input']>;
  name_pattern?: InputMaybe<Scalars['String']['input']>;
  templateids?: InputMaybe<Array<InputMaybe<Scalars['String']['input']>>>;
}


export interface QueryLoginArgs {
  password: Scalars['String']['input'];
  username: Scalars['String']['input'];
}


export interface QueryUserPermissionsArgs {
  objectNames?: InputMaybe<Array<Scalars['String']['input']>>;
}
// ...
export interface UserRightsInput {
  userGroups?: InputMaybe<Array<UserGroupInput>>;
  userRoles?: InputMaybe<Array<UserRoleInput>>;
}
// ...
export type ApiErrorResolvers<ContextType = any, ParentType extends ResolversParentTypes['ApiError'] = ResolversParentTypes['ApiError']> = {
  args?: Resolver<Maybe<ResolversTypes['JSONObject']>, ParentType, ContextType>;
  code?: Resolver<Maybe<ResolversTypes['Int']>, ParentType, ContextType>;
  data?: Resolver<Maybe<ResolversTypes['JSONObject']>, ParentType, ContextType>;
  message?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  path?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type CreateHostGroupResponseResolvers<ContextType = any, ParentType extends ResolversParentTypes['CreateHostGroupResponse'] = ResolversParentTypes['CreateHostGroupResponse']> = {
  error?: Resolver<Maybe<ResolversTypes['ApiError']>, ParentType, ContextType>;
  groupName?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  groupid?: Resolver<Maybe<ResolversTypes['Int']>, ParentType, ContextType>;
  message?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type CreateHostResponseResolvers<ContextType = any, ParentType extends ResolversParentTypes['CreateHostResponse'] = ResolversParentTypes['CreateHostResponse']> = {
  deviceKey?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  error?: Resolver<Maybe<ResolversTypes['ApiError']>, ParentType, ContextType>;
  hostid?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  message?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export interface DateTimeScalarConfig extends GraphQLScalarTypeConfig<ResolversTypes['DateTime'], any> {
  name: 'DateTime';
}

export type DeviceResolvers<ContextType = any, ParentType extends ResolversParentTypes['Device'] = ResolversParentTypes['Device']> = {
  __resolveType: TypeResolveFn<'GenericDevice', ParentType, ContextType>;
  deviceType?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  host?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  hostgroups?: Resolver<Maybe<Array<ResolversTypes['HostGroup']>>, ParentType, ContextType>;
  hostid?: Resolver<ResolversTypes['ID'], ParentType, ContextType>;
  name?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  state?: Resolver<Maybe<ResolversTypes['DeviceState']>, ParentType, ContextType>;
  tags?: Resolver<Maybe<ResolversTypes['JSONObject']>, ParentType, ContextType>;
};

export type DeviceCommunicationTypeResolvers = EnumResolverSignature<{ DATABASE_MONITOR?: any, DEPENDANT_ITEM?: any, HTTP_AGENT?: any, IPMI_AGENT?: any, JMX_AGENT?: any, SIMPLE_CHECK?: any, SIMULATOR_CALCULATED?: any, SIMULATOR_JAVASCRIPT?: any, SNMP_AGENT?: any, SNMP_TRAP?: any, ZABBIX_AGENT?: any, ZABBIX_AGENT_ACTIVE?: any, ZABBIX_INTERNAL_ITEM?: any, ZABBIX_TRAP?: any }, ResolversTypes['DeviceCommunicationType']>;

export type DeviceStateResolvers<ContextType = any, ParentType extends ResolversParentTypes['DeviceState'] = ResolversParentTypes['DeviceState']> = {
  __resolveType: TypeResolveFn<'GenericDeviceState', ParentType, ContextType>;
  operational?: Resolver<Maybe<ResolversTypes['OperationalDeviceData']>, ParentType, ContextType>;
};

export type DeviceStatusResolvers = EnumResolverSignature<{ DISABLED?: any, ENABLED?: any }, ResolversTypes['DeviceStatus']>;

export type ErrorResolvers<ContextType = any, ParentType extends ResolversParentTypes['Error'] = ResolversParentTypes['Error']> = {
  __resolveType: TypeResolveFn<'ApiError', ParentType, ContextType>;
  code?: Resolver<Maybe<ResolversTypes['Int']>, ParentType, ContextType>;
  data?: Resolver<Maybe<ResolversTypes['JSONObject']>, ParentType, ContextType>;
  message?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
};

export type ErrorPayloadResolvers<ContextType = any, ParentType extends ResolversParentTypes['ErrorPayload'] = ResolversParentTypes['ErrorPayload']> = {
  additionalInfo?: Resolver<Maybe<ResolversTypes['JSONObject']>, ParentType, ContextType>;
  code?: Resolver<ResolversTypes['Int'], ParentType, ContextType>;
  message?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};
// ...
export type Resolvers<ContextType = any> = {
  ApiError?: ApiErrorResolvers<ContextType>;
  CreateHostGroupResponse?: CreateHostGroupResponseResolvers<ContextType>;
  CreateHostResponse?: CreateHostResponseResolvers<ContextType>;
  DateTime?: GraphQLScalarType;
  Device?: DeviceResolvers<ContextType>;
  DeviceCommunicationType?: DeviceCommunicationTypeResolvers;
  DeviceState?: DeviceStateResolvers<ContextType>;
  DeviceStatus?: DeviceStatusResolvers;
  Error?: ErrorResolvers<ContextType>;
  ErrorPayload?: ErrorPayloadResolvers<ContextType>;
  GenericDevice?: GenericDeviceResolvers<ContextType>;
  GenericDeviceState?: GenericDeviceStateResolvers<ContextType>;
  GpsPosition?: GpsPositionResolvers<ContextType>;
  HistoryExportResponse?: HistoryExportResponseResolvers<ContextType>;
  Host?: HostResolvers<ContextType>;
  HostGroup?: HostGroupResolvers<ContextType>;
  HostTypeMeta?: HostTypeMetaResolvers<ContextType>;
  ImportUserRightResult?: ImportUserRightResultResolvers<ContextType>;
  ImportUserRightsResult?: ImportUserRightsResultResolvers<ContextType>;
  Inventory?: InventoryResolvers<ContextType>;
  Item?: ItemResolvers<ContextType>;
  JSONObject?: GraphQLScalarType;
  Location?: LocationResolvers<ContextType>;
  Mutation?: MutationResolvers<ContextType>;
  OperationalDeviceData?: OperationalDeviceDataResolvers<ContextType>;
  Permission?: PermissionResolvers;
  Query?: QueryResolvers<ContextType>;
  StorageItemType?: StorageItemTypeResolvers;
  Template?: TemplateResolvers<ContextType>;
  Time?: GraphQLScalarType;
  UserGroup?: UserGroupResolvers<ContextType>;
  UserPermission?: UserPermissionResolvers<ContextType>;
  UserRights?: UserRightsResolvers<ContextType>;
  UserRole?: UserRoleResolvers<ContextType>;
  UserRoleModule?: UserRoleModuleResolvers<ContextType>;
  UserRoleRule?: UserRoleRuleResolvers<ContextType>;
  UserRoleRules?: UserRoleRulesResolvers<ContextType>;
  ZabbixCreateResponse?: ZabbixCreateResponseResolvers<ContextType>;
  ZabbixGroupRight?: ZabbixGroupRightResolvers<ContextType>;
  ZabbixHost?: ZabbixHostResolvers<ContextType>;
  ZabbixItem?: ZabbixItemResolvers<ContextType>;
};